third_party/WebKit/LayoutTests/http/tests/notifications/notification-sandbox-permission.html

<!DOCTYPE html>
<html>
    <body>
        <p>
            Tests that the &lt;iframe sandbox&gt; attribute influences whether
            notification permissions will be inherit by an iframe.
        </p>
        <div id="console"></div>
        <script>
            if (window.testRunner) {
                testRunner.setPermission('notifications', 'granted', location.origin, location.origin);
                testRunner.dumpAsText();
                testRunner.dumpChildFrames();
            }
        </script>

        <h1>Normal iframe. Should inherit permissions.</h1>
        <iframe src="resources/display-permission.html"></iframe>

        <h1>Sandboxed iframe with allow-same-origin. Should inherit permissions.</h1>
        <iframe src="resources/display-permission.html" sandbox="allow-scripts allow-same-origin"></iframe>

        <h1>Sandboxed iframe without allow-same-origin. Should not inherit permissions.</h1>
        <iframe src="resources/display-permission.html" sandbox="allow-scripts"></iframe>
    </body>
</html>
