third_party/blink/web_tests/http/tests/notifications/notification-sandbox-permission-expected.txt

Tests that the <iframe sandbox> attribute influences whether notification permissions will be inherit by an iframe.

Normal iframe. Should inherit permissions.
Sandboxed iframe with allow-same-origin. Should inherit permissions.
Sandboxed iframe without allow-same-origin. Should not inherit permissions.

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
granted

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
granted

--------
Frame: '<!--framePath //<!--frame2-->-->'
--------
denied
